Impact handling in new rule

+++ src/main/java/com/facilio/bmsconsole/context/BaseEventContext.java
@@ -23,15 +23,16 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 public class BaseEventContext extends ModuleBaseWithCustomFields {
 	private static final long serialVersionUID = 1L;
 	
-	private long cost;
+	private double cost;
 
-	public long getCost() {
+	public double getCost() {
 		return cost;
 	}
 
-	public void setCost(long cost) {
+	public void setCost(double cost) {
 		this.cost = cost;
 	}
+
 
 	private String description;
 	private boolean superCalled = false;
